* changed in bench/spec_test.dat: yaw limit to pitch limit        * changed in bench/*qq files: yaw to pitch and vice versa        * fixed bug in computation of standard deviations in ec_ncdf.f: were set to -9999 due to check on number of samples        * converted tabs to spaces in some files

+++ libsrc/ec_corr.f
@@ -335,7 +335,7 @@ C-- Start Simpson's rule numerical integration -----------------------
 C     From -5 to log(5) in NumInt steps
 C
       I3=0
-      DO I1=1,0.5*(NumInt+1)
+      DO I1=1,INT(0.5*(NumInt+1))
         DO I2=2,4,2
           I3=I3+1
           IF (I3.GT.NumInt) GOTO 200
